Move all of Test/Dummy devices into microscope.testsuite
